--- bzt/modules/blazemeter/cloud_const.py
@@ -1,22 +1,0 @@
-"""
-Module for reporting into http://www.blazemeter.com/ service
-
-Copyright 2015 BlazeMeter Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-
-LOC = "locations"
-LOC_WEIGHTED = "locations-weighted"
-DEDICATED_IPS = "dedicated-ips"
-NOTE_SIZE_LIMIT = 2048
